requirements.txt
CTkListbox==1.4
CTkMessagebox==2.5
customtkinter==5.2.2
entsoe_py==0.6.7
matplotlib==3.7.1
nordpool==0.3.3
numpy==1.23.5
opencv_python==4.7.0.72
pandas==2.2.2
Pillow==10.3.0
pvlib==0.10.4
python_dateutil==2.8.2
requests==2.32.3
retry_requests==2.0.0
